move currentlySupportedLangs out of folder
