+++ config/webpack.common.config.js
@@ -6,7 +6,7 @@ module.exports = {
   entry: {
     assets: path.resolve(__dirname, '../src/index.jsx'),
     accessibilityPolicy: path.resolve(__dirname, '../src/accessibilityIndex.jsx'),
-    i18nMessages: path.resolve(__dirname, '../src/data/i18n/locales/json/currentlySupportedLangs.jsx'),
+    i18nMessages: path.resolve(__dirname, '../src/data/i18n/locales/currentlySupportedLangs.jsx'),
   },
   output: {
     filename: '[name].js',
